requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
django-filter>=23.2
drf-yasg>=1.21.6
python-dotenv>=1.0.0
Pillow>=9.5.0
psycopg2-binary>=2.9.6
djangorestframework-simplejwt>=5.2.2
gunicorn>=20.1.0
whitenoise>=6.5.0
